Solution to Check Profit or Loss

+++ Check Profit or Loss@@6788dd63490637cfc44e1b98/code.c
@@ -2,8 +2,8 @@
 # include <stdio.h>
 
 int main() {
-    int cost_price, selling_price;
-    scanf("%d %d", &cost_price, &selling_price);
+    float cost_price, selling_price;
+    scanf("%f %f", &cost_price, &selling_price);
     
     if (cost_price < selling_price) {
         printf("Profit");
